# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Reliable Checkout and Customer Rewards API",
    "version": "1.0.0",
    "description": "Use **Try it out** and **Execute** below to call this running service. No API client installation or credentials are needed.\n\n### Quick walkthrough\n1. Create a customer and copy its returned id.\n2. List products; the seeded Notebook ID is already provided as an item-path example.\n3. Create a cart with your real customer ID and copy its returned id.\n4. Add an item using PUT and your real cart ID.\n5. Checkout with a unique Idempotency-Key and the default empty body.\n6. Execute the identical checkout again to verify durable replay, then retrieve its order.\n7. Inspect customer coupons and the admin report. Use an owned coupon code on a later cart to redeem it.\n\nExample UUIDs are placeholders: replace customer/cart/order IDs with values returned by your requests. Use a new checkout key for each new cart, retaining the same key and coupon on retries.\n\nMilestones use the environment policy, default n=5 and x=10. Every fifth store-wide order automatically earns a customer-specific coupon. Admin generation normally returns 409 because it only repairs missing historical rewards.\n\nPrices are exact decimal INR strings. Carts use live prices without reservations. Orders are immutable snapshots. Authentication is out of scope: customer IDs are trusted inputs and Administration operations are administrative."
  },
  "servers": [
    {
...
    }
...
  "paths": {
...
    "/customers": {
      "post": {
        "tags": ["Customers"],
        "operationId": "createCustomer",
        "summary": "Create a customer",
        "description": "Returns the customer ID to use when creating carts. Retrying creates another customer.",
        "responses": {
          "201": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "413": {
            "$ref": "#/components/responses/BodyTooLarge"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/DatabaseBusy"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomer"
              }
            }
          }
        }
      }
    },
    "/products": {
      "get": {
        "tags": ["Products"],
        "operationId": "listProducts",
        "summary": "List products and inventory",
        "description": "Returns all seeded products, including zero-stock items. Prices are current INR decimal strings.",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                },
                "example": [
                  {
                    "id": "00000000-0000-4000-8000-000000000001",
                    "name": "Notebook",
                    "inventory": 100,
                    "unitPrice": "200.34",
                    "currency": "INR"
                  }
                ]
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/carts": {
      "post": {
        "tags": ["Carts"],
        "operationId": "createCart",
        "summary": "Create a customer-owned cart",
        "description": "Use a real customer ID returned by Create a customer. The new cart is empty. Retrying creates another cart.",
        "responses": {
          "201": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                },
                "example": {
                  "id": "22222222-2222-4222-8222-222222222222",
                  "customerId": "11111111-1111-4111-8111-111111111111",
                  "status": "open",
                  "createdAt": "2026-10-06T14:00:00.000Z",
                  "currency": "INR",
                  "items": [],
                  "subtotal": "0.00",
                  "discount": "0.00",
                  "total": "0.00",
                  "orderId": null
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "description": "Customer does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "CUSTOMER_NOT_FOUND": {
                    "value": {
                      "error": {
                        "code": "CUSTOMER_NOT_FOUND",
                        "message": "Customer does not exist."
                      }
                    }
                  }
                }
              }
            }
          },
          "413": {
            "$ref": "#/components/responses/BodyTooLarge"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/DatabaseBusy"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCart"
              }
            }
          }
        }
      }
    },
    "/carts/{cartId}": {
      "get": {
        "tags": ["Carts"],
        "operationId": "getCart",
        "summary": "View a cart",
        "description": "Shows live prices and current availability. Checked-out carts include orderId; retrieve that order for actual purchase snapshots.",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "description": "Cart does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "CART_NOT_FOUND": {
                    "value": {
                      "error": {
                        "code": "CART_NOT_FOUND",
                        "message": "Cart does not exist."
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/CartId"
          }
        ]
      }
    },
...
    "/orders/{orderId}": {
      "get": {
        "tags": ["Orders"],
        "operationId": "getOrder",
        "summary": "Retrieve an immutable order",
        "description": "Product names, prices and totals reflect purchase time even if the catalog changes.",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "description": "Order does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "ORDER_NOT_FOUND": {
                    "value": {
                      "error": {
                        "code": "ORDER_NOT_FOUND",
                        "message": "Order does not exist."
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/OrderId"
          }
        ]
      }
    },
...
    "/admin/products": {
      "post": {
        "tags": ["Administration"],
        "operationId": "createProduct",
        "summary": "Create a product",
        "description": "Administrative. Creates a generated UUID product with exact INR unit price and nonnegative stock. All three fields are required; currency is always INR. Retrying creation can create another product.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProduct"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product with normalized decimal INR price.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "413": {
            "$ref": "#/components/responses/BodyTooLarge"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/DatabaseBusy"
          }
        }
      }
    },
    "/admin/products/{productId}": {
      "patch": {
        "tags": ["Administration"],
        "operationId": "updateProduct",
        "summary": "Update product name, price or stock",
        "description": "Administrative partial update. Supply at least one field; omitted fields are preserved and ID cannot change. Inventory replaces the absolute current available count. Updates serialize with checkout. Existing carts use the updated price and availability, while previous orders and revenue retain their purchase snapshots.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProductId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProduct"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product with normalized decimal INR price.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "description": "Product does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "PRODUCT_NOT_FOUND",
                    "message": "Product not found"
                  }
                }
              }
            }
          },
          "413": {
            "$ref": "#/components/responses/BodyTooLarge"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/DatabaseBusy"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Money": {
        "type": "string",
        "pattern": "^(0|[1-9][0-9]*)\\.[0-9]{2}$",
        "description": "Exact INR decimal string with two fractional digits; never a floating-point JSON number.",
        "example": "200.34"
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string",
                "example": "INSUFFICIENT_INVENTORY"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "description": "Optional domain details or array of validation issues; shape depends on error code."
              }
            }
          }
        }
      },
      "Customer": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "name", "createdAt"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "example": {
          "id": "11111111-1111-4111-8111-111111111111",
          "name": "Asha",
          "createdAt": "2026-10-06T14:00:00.000Z"
        }
      },
      "Product": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "name", "inventory", "unitPrice", "currency"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "inventory": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "unitPrice": {
            "$ref": "#/components/schemas/Money"
          },
          "currency": {
            "type": "string",
            "enum": ["INR"]
          }
        },
        "example": {
          "id": "00000000-0000-4000-8000-000000000001",
          "name": "Notebook",
          "inventory": 100,
          "unitPrice": "200.34",
          "currency": "INR"
        }
      },
      "CartItem": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "productId",
          "name",
          "quantity",
          "unitPrice",
          "lineTotal",
          "availableInventory",
          "available"
        ],
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "unitPrice": {
            "$ref": "#/components/schemas/Money"
          },
          "lineTotal": {
            "$ref": "#/components/schemas/Money"
          },
          "availableInventory": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "available": {
            "type": "boolean"
          }
        }
      },
      "Cart": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "customerId",
          "status",
          "createdAt",
          "currency",
          "items",
          "subtotal",
          "discount",
          "total",
          "orderId"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": ["open", "checked_out"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "currency": {
            "type": "string",
            "enum": ["INR"]
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItem"
            }
          },
          "subtotal": {
            "$ref": "#/components/schemas/Money"
          },
          "discount": {
            "$ref": "#/components/schemas/Money"
          },
          "total": {
            "$ref": "#/components/schemas/Money"
          },
          "orderId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "example": {
          "id": "22222222-2222-4222-8222-222222222222",
          "customerId": "11111111-1111-4111-8111-111111111111",
          "status": "open",
          "createdAt": "2026-10-06T14:00:00.000Z",
          "currency": "INR",
          "items": [
            {
              "productId": "00000000-0000-4000-8000-000000000001",
              "name": "Notebook",
              "quantity": 1,
              "unitPrice": "200.34",
              "lineTotal": "200.34",
              "availableInventory": 100,
              "available": true
            }
          ],
          "subtotal": "200.34",
          "discount": "0.00",
          "total": "200.34",
          "orderId": null
        }
      },
      "OrderItem": {
        "type": "object",
        "additionalProperties": false,
        "required": ["productId", "name", "quantity", "unitPrice", "lineTotal"],
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "unitPrice": {
            "$ref": "#/components/schemas/Money"
          },
          "lineTotal": {
            "$ref": "#/components/schemas/Money"
          }
        }
      },
...
      "Order": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "cartId",
          "customerId",
          "ordinal",
          "currency",
          "createdAt",
          "items",
          "subtotal",
          "discount",
          "total",
          "appliedCoupon"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "cartId": {
            "type": "string",
            "format": "uuid"
          },
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "ordinal": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "currency": {
            "type": "string",
            "enum": ["INR"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            },
            "minItems": 1
          },
          "subtotal": {
            "$ref": "#/components/schemas/Money"
          },
          "discount": {
            "$ref": "#/components/schemas/Money"
          },
          "total": {
            "$ref": "#/components/schemas/Money"
          },
          "appliedCoupon": {
            "$ref": "#/components/schemas/NullableAppliedCoupon"
          }
        },
        "example": {
          "id": "33333333-3333-4333-8333-333333333333",
          "cartId": "22222222-2222-4222-8222-222222222222",
          "customerId": "11111111-1111-4111-8111-111111111111",
          "ordinal": 1,
          "currency": "INR",
          "createdAt": "2026-10-06T14:00:00.000Z",
          "items": [
            {
              "productId": "00000000-0000-4000-8000-000000000001",
              "name": "Notebook",
              "quantity": 1,
              "unitPrice": "200.34",
              "lineTotal": "200.34"
            }
          ],
          "subtotal": "200.34",
          "discount": "0.00",
          "total": "200.34",
          "appliedCoupon": null
        }
      },
...
      "CreateCustomer": {
        "type": "object",
        "additionalProperties": false,
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Trimmed before validating length."
          }
        },
        "example": {
          "name": "Asha"
        }
      },
      "CreateCart": {
        "type": "object",
        "additionalProperties": false,
        "required": ["customerId"],
        "properties": {
          "customerId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "example": {
          "customerId": "11111111-1111-4111-8111-111111111111"
        }
      },
...
      "NullableAppliedCoupon": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "discountPercent"],
        "properties": {
          "code": {
            "type": "string"
          },
          "discountPercent": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          }
        },
        "nullable": true,
        "example": null
      },
      "CreateProduct": {
        "type": "object",
        "additionalProperties": false,
        "required": ["name", "unitPrice", "inventory"],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Trimmed before validating length."
          },
          "unitPrice": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32,
            "pattern": "^(0|[1-9][0-9]*)(\\.[0-9]{1,2})?$",
            "description": "Exact nonnegative INR decimal string, at most two fractional digits, up to 90071992547409.91.",
            "example": "200.34"
          },
          "inventory": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Absolute available stock count, not a delta."
          }
        },
        "example": {
          "name": "Desk Lamp",
          "unitPrice": "200.34",
          "inventory": 10
        }
      },
      "UpdateProduct": {
        "type": "object",
        "additionalProperties": false,
        "minProperties": 1,
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Trimmed before validating length."
          },
          "unitPrice": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32,
            "pattern": "^(0|[1-9][0-9]*)(\\.[0-9]{1,2})?$",
            "description": "Exact nonnegative INR decimal string, at most two fractional digits, up to 90071992547409.91.",
            "example": "200.34"
          },
          "inventory": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Absolute available stock count, not a delta."
          }
        },
        "example": {
          "unitPrice": "249.99",
          "inventory": 5
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed UUID, body, key, or unsupported numeric range.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "VALIDATION_ERROR": {
                "value": {
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Malformed UUID, body, key, or unsupported numeric range."
                  }
                }
              },
              "INVALID_JSON": {
                "value": {
                  "error": {
                    "code": "INVALID_JSON",
                    "message": "Malformed UUID, body, key, or unsupported numeric range."
                  }
                }
              },
              "AMOUNT_OUT_OF_RANGE": {
                "value": {
                  "error": {
                    "code": "AMOUNT_OUT_OF_RANGE",
                    "message": "Malformed UUID, body, key, or unsupported numeric range."
                  }
                }
              },
              "INVALID_MONEY": {
                "value": {
                  "error": {
                    "code": "INVALID_MONEY",
                    "message": "Use a nonnegative decimal string with at most two fractional digits"
                  }
                }
              }
            }
          }
        }
      },
      "InternalError": {
        "description": "Unexpected failure; logged server-side with sanitized public details.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "INTERNAL_ERROR": {
                "value": {
                  "error": {
                    "code": "INTERNAL_ERROR",
                    "message": "Unexpected failure; logged server-side with sanitized public details."
                  }
                }
              }
            }
          }
        }
      },
      "BodyTooLarge": {
        "description": "JSON request body exceeds 16kb.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "BODY_TOO_LARGE": {
                "value": {
                  "error": {
                    "code": "BODY_TOO_LARGE",
                    "message": "JSON request body exceeds 16kb."
                  }
                }
              }
            }
          }
        }
      },
      "DatabaseBusy": {
        "description": "Database write-lock timeout. Retry; for checkout keep the same idempotency key.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "DATABASE_BUSY": {
                "value": {
                  "error": {
                    "code": "DATABASE_BUSY",
                    "message": "Database write-lock timeout. Retry; for checkout keep the same idempotency key."
                  }
                }
              }
            }
          }
        },
        "headers": {
          "Retry-After": {
            "description": "Suggested delay in seconds.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        }
      }
    },
    "parameters": {
      "CartId": {
        "name": "cartId",
        "in": "path",
        "required": true,
        "description": "Replace with a cart ID returned by Create a cart.",
        "schema": {
          "type": "string",
          "format": "uuid",
          "example": "22222222-2222-4222-8222-222222222222"
        }
      },
      "ProductId": {
        "name": "productId",
        "in": "path",
        "required": true,
        "description": "Seeded Notebook ID, or another ID returned by List products.",
        "schema": {
          "type": "string",
          "format": "uuid",
          "example": "00000000-0000-4000-8000-000000000001"
        }
      },
      "OrderId": {
        "name": "orderId",
        "in": "path",
        "required": true,
        "description": "Replace with an order ID returned by Checkout.",
        "schema": {
          "type": "string",
          "format": "uuid",
          "example": "33333333-3333-4333-8333-333333333333"
        }
      },
...
    }
  }
}